backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2025.3
SECRET_KEY=your-super-secret-key-change-this-later
DEBUG=True
DB_NAME=dietplanner_db
DB_USER=postgres
DB_PASSWORD=your_postgres_password
DB_HOST=localhost
DB_PORT=5432
GEMINI_API_KEY=your-gemini-api-key-here
pydantic>=2.0